# api/openapi.yaml
openapi: 3.0.3
info:
  title: Ecommerce API
  version: 0.2.0
  description: API contract for backend and frontend type generation.
servers:
  - url: http://localhost:3000
security:
  - cookieAuth: []
  - bearerAuth: []
...
paths:
...
  /api/v1/auth/oidc/login:
    get:
      tags: [auth]
      operationId: oidcLogin
      security: []
      parameters:
        - in: query
          name: redirect
          schema:
            type: string
      responses:
        "302":
          description: Redirect to provider
  /api/v1/auth/oidc/callback:
    get:
      tags: [auth]
      operationId: oidcCallback
      security: []
      parameters:
        - in: query
          name: code
          schema:
            type: string
        - in: query
          name: state
          schema:
            type: string
        - in: query
          name: format
          schema:
            type: string
            enum: [json]
      responses:
        "200":
          description: Callback as JSON
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "302":
          description: Callback as redirect
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/products:
    get:
      tags: [products]
      operationId: listProducts
      security: []
      summary: List products
      parameters:
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: min_price
          schema:
            type: number
            format: double
        - in: query
          name: max_price
          schema:
            type: number
            format: double
        - in: query
          name: brand_slug
          schema:
            type: string
        - in: query
          name: has_variant_stock
          schema:
            type: boolean
        - in: query
          name: attribute
          style: deepObject
          explode: true
          schema:
            type: object
            additionalProperties:
              type: string
        - in: query
          name: sort
          schema:
            type: string
            enum: [price, name, created_at]
        - in: query
          name: order
          schema:
            type: string
            enum: [asc, desc]
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Paginated products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductPage"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/v1/products/{id}:
    get:
      tags: [products]
      operationId: getProduct
      security: []
      summary: Get product by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/storefront:
    get:
      tags: [storefront]
      operationId: getStorefrontSettings
      security: []
      responses:
        "200":
          description: Storefront settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StorefrontSettingsResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/me/:
    get:
      tags: [profile]
      operationId: getProfile
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      tags: [profile]
      operationId: updateProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/me/profile-photo:
    post:
      tags: [media]
      operationId: setProfilePhoto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [media_id]
              properties:
                media_id:
                  type: string
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: Too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [media]
      operationId: deleteProfilePhoto
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /api/v1/me/payment-methods:
    get:
      tags: [profile]
      operationId: listSavedPaymentMethods
      responses:
        "200":
          description: Saved payment methods
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavedPaymentMethod"
    post:
      tags: [profile]
      operationId: createSavedPaymentMethod
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSavedPaymentMethodRequest"
      responses:
        "201":
          description: Created payment method
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedPaymentMethod"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/me/payment-methods/{id}:
    delete:
      tags: [profile]
      operationId: deleteSavedPaymentMethod
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/me/payment-methods/{id}/default:
    patch:
      tags: [profile]
      operationId: setDefaultPaymentMethod
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Updated payment method
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedPaymentMethod"
  /api/v1/me/addresses:
    get:
      tags: [profile]
      operationId: listSavedAddresses
      responses:
        "200":
          description: Saved addresses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavedAddress"
    post:
      tags: [profile]
      operationId: createSavedAddress
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSavedAddressRequest"
      responses:
        "201":
          description: Created address
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedAddress"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/me/addresses/{id}:
    delete:
      tags: [profile]
      operationId: deleteSavedAddress
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/me/addresses/{id}/default:
    patch:
      tags: [profile]
      operationId: setDefaultAddress
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Updated address
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedAddress"
  /api/v1/me/cart:
    get:
      tags: [cart]
      operationId: getCart
      responses:
        "200":
          description: Cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    post:
      tags: [cart]
      operationId: addCartItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddCartItemRequest"
      responses:
        "200":
          description: Cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /api/v1/checkout/cart:
    get:
      tags: [checkout]
      operationId: getCheckoutCart
      responses:
        "200":
          description: Cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
        "403":
          description: Guest checkout disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/checkout/cart/summary:
    get:
      tags: [checkout]
      operationId: getCheckoutCartSummary
      responses:
        "200":
          description: Cart summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutCartSummary"
        "403":
          description: Guest checkout disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/checkout/cart/items:
    post:
      tags: [checkout]
      operationId: addCheckoutCartItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddCartItemRequest"
      responses:
        "200":
          description: Cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
        "403":
          description: Guest checkout disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session_token
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    AuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"
    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
        currency:
          type: string
          minLength: 3
          maxLength: 3
        profile_photo_url:
          type: string
...
    ProductPage:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        pagination:
          $ref: "#/components/schemas/Pagination"
...
    CheckoutCartSummary:
      type: object
      required: [item_count]
      properties:
        item_count:
          type: integer
          minimum: 0
...
    StorefrontSettingsResponse:
      type: object
      required: [settings, updated_at, has_draft_changes, published_updated_at]
      properties:
        settings:
          $ref: "#/components/schemas/StorefrontSettings"
        updated_at:
          type: string
          format: date-time
        has_draft_changes:
          type: boolean
        draft_updated_at:
          type: string
          format: date-time
          nullable: true
        published_updated_at:
          type: string
          format: date-time